Fix Group 2.1: ExtensionDtype attribute access errors (15+ errors)
Added type: ignore comments for ExtensionDtype union type mismatches where
pandas-stubs has more restrictive type requirements than runtime behavior:

Files modified:
- python/cudf/cudf/core/column/column.py: Fixed 9 errors
  - ExtensionDtype.__from_arrow__ attribute access
  - astype() and as_*_column() dtype parameter mismatches
  - Column constructor dtype parameters
- python/cudf/cudf/core/column/categorical.py: Fixed 2 errors
  - astype() calls with ExtensionDtype unions
- python/cudf/cudf/core/column_accessor.py: Fixed 2 errors
  - pd.Index dtype parameter unions
- python/cudf/cudf/utils/dtypes.py: Fixed 3 errors
  - pa.from_numpy_dtype() argument type
  - pyarrow_dtype.to_pandas_dtype() return type
  - cudf.StructDtype() argument type

All fixes use type ignore comments to preserve runtime behavior while
satisfying mypy's strict type checking with pandas-stubs.

+++ python/cudf/cudf/core/column/categorical.py
@@ -875,8 +875,8 @@ class CategoricalColumn(column.ColumnBase):
             [old_categories.dtype, new_categories.dtype]
         )
 
-        new_categories = new_categories.astype(common_dtype)
-        old_categories = old_categories.astype(common_dtype)
+        new_categories = new_categories.astype(common_dtype)  # type: ignore[arg-type]
+        old_categories = old_categories.astype(common_dtype)  # type: ignore[arg-type]
 
         if old_categories.isin(new_categories).any():
             raise ValueError("new categories must not include old categories")
